Add example and fix typo for readme of regex versioning module
This adds an example for linuxserver/tautulli docker image to show that for the fourth version part (the ls number), build should be used. This differs from the current example of linuxserver/openssh-server which the ls number is the fifth version part so versioning should be used. Closes discussion #33270. Also fixes typo in the doc.