feat(azure): add support for rebase-merge
Azure DevOps has support for semi-linear rebase (rebase with merge
commit, see [1]). Our branch policies enforce this option, which is
incompatible with any of the `automergeStrategy` options currently
available.

In fact, when this option is set to auto, Renovate Bot will actually
select semi-linear rebasing if that is the only option. But if there
are other (allowed) options, then there is nothing to enforce it.

This adds the `rebase-merge` option, which is different from using
`rebase` (again, see [1]).

Based on closer inspection, I see that Gitea will do a rebase with
merge when the option is `rebase`. I did not adapt this, because that
would break current configurations. Also, I only have the option to
test this with Azure DevOps.

[1]: https://devblogs.microsoft.com/devops/pull-requests-with-rebase/