refactor(bazel-module): use bazel syntactic concepts in parser
Instead of using a general purpose "record" type, we now parse into
rules and extension tags (see comments in the relevant files for a
bazel primer of what these are).

This change allows us to de-duplicate some of the parsing
functionality and hopefully makes it easier to add support for more
bazel packages. Notably, it should be possible to add support for a
new bazel rule without having to understand parsing.

As an added benefit, this centralizes the way we recognize extensions,
so drift between different types of extensions is less likely.

This refactor was triggered by #33670. To fix that, we need to keep
the original syntactic strings during parsing. This is made easier by
this cleanup.