test(bazel-module): remove Ctx success case tests
The success paths of Ctx are already fully tested via the parser
tests (parser/index.spec.ts).

This is triggered by #33670, for which we need to adjust parsing to
retain the raw original strings so we can provide them as
`replaceString` in the `PackageDepencency`.

This in turn requires adjustments to the context that become
prohibitively hard to test at such a low level.

Since (as coverage shows), these tests are not necessary, we simply
remove them.