fix(nix): urldecode gitlab subgroups
Do to a bug in nix one needs to uri encode the owner of a gitlab repository
that resides in a subgroups, in this case the `/` as `%20F`. This
however breaks the `git ls-remotes` which renovate runs to discover if
the input can be updated.

This commit passes the owner attribute through `decodeURIComponent` to ensure
that `git` is unware of this workaround on nixs side.

Nix PR: https://github.com/NixOS/nix/pull/9163