feat: Support multi-version python-version files
If a `.python-version` file contains multiple
lines, each line, that contains a valid version
string is returned.
A version is here an arbitrary string, that
does not start with a hash-symbol.
A line may contain comments starting with hashes.
In this case, the hash symbol and the trailing
characters will be ignored.
Empty lines will also be ignored.

This is especially useful, when environments like
tox shall test multiple python versions at once.

The position of the version within the file must
now be taken into account.

If a version has major, minor or patch version
parts, the specific bump version property is
set.

Closes #31850