chore(gomod): Specs for go.mod directives unrelated to dependencies
This adds a few unit tests for go.mod directives that are not related to
dependencies: `godebug`, `retract` and single-line `exclude`. All of
those should be ignored.

> The `exclude` directive prevents a module version from being loaded by
> the `go` command.
> https://go.dev/ref/mod#go-mod-file-exclude

Apparently they can be used to e.g. block a specific version of a
dependency which has a known vulnerability. Simply updating the
`exclude` section doesn't make the most sense as someone wanted to
exclude a specific version of a dependency.

Note: This commit doesn't fix the behaviour for multi-line excludes, it
just increases test coverage to show that single-line excludes are
already ignored.

References:
- https://go.dev/doc/modules/gomod-ref