fix(gomod): Exclude multi-line `exclude` directives from extraction
With https://github.com/renovatebot/renovate/pull/28852, suddenly
multiline `exclude` directives have been renovated.

Given that single-line `exclude` directives are not renovated, this just
aligns the behaviour between the two. In a future iteration the
`exclude`s could be parsed in order to change the behaviour for renovate
with regards to the dependency updates, but it is unclear what that
would entail.

So simply skipping them could be good enough.

Side-note: The regex based parsing seems to be really tricky for a lock
file which might contain multi line blocks. One should consider
rewriting it in a proper loop.

References:
- https://github.com/renovatebot/renovate/discussions/31905