feat(config)!: add minor version to default branchTopic to support separateMultipleMinor (#35606)
Changes the default branchTopic to include the minor version when separateMultipleMinor is enabled, ensuring that separateMultipleMinor works properly when hashedBranchLength is also set.

BREAKING CHANGE: This will be a breaking change for any users that have customized their branchTopic and use separateMultipleMinor=true. In those cases, the minor version version will no longer be included in the branch names until they make a similar modification to their branchTopic. For all other users (who haven't set branchTopic or separateMultipleMinor) the behavior should be unchanged.