Add support for uv workspaces
In uv, [workspace repos](https://docs.astral.sh/uv/concepts/projects/workspaces/)
consist of a "root package" along with a set of other packages that are all
managed with a single environment, and therefore a single lockfile.
Currently, renovate locates the `uv.lock` file for a given project by
concatenating the directory of the `pyproject.toml` file with the filename
`uv.lock`. This is accurate for non-workspaces.

This PR takes a relatively simple approach, using an existing utility function to
search for a `uv.lock` file up the file tree, stopping at the first one found.
Missing lockfiles are treated identically to missing lockfiles in the "sibling-only"
/ single-project case.

For re-locking, no change is needed; uv itself will handle the lockfile resolution
via `uv lock`, even if it is run in a subdirectory.