build(ls-lint): ignore Git worktrees from linting
When working with multiple branches of a project, it is common to use
Git Worktrees[0].

However, this can lead to Husky failing pre-commit checks with i.e.

    .worktrees/.../.github/pull_request_template.md failed for `.md` rules: kebabcase
    .worktrees/.../CODE_OF_CONDUCT.md failed for `.md` rules: kebabcase
    .worktrees/.../SECURITY.md failed for `.md` rules: kebabcase
    husky - pre-commit script failed (code 1)

We can therefore add - an opinionated - directory for these worktrees to
be set in, where they won't be hit by Husky and `ls-lint`'s linting.

[0]: https://git-scm.com/docs/git-worktree