fix(jsonata): Eval later queries despite non-match
The jsonata custom manager can be provided with multiple queries via the
matchStrings configuration option.

Before this commit, if any of those doesn't match for a file, all later
querites are skipped. This doesn't make sense, since the queries can be
completely independent. As an example, one might want to look for image
names in several places of every CI pipeline. Not matching in one place
of one pipeline shouldn't cause the other places to be ignored in that
file.

This commit thus changes that behavior to continue with the other
queries in the same file if one doesn't match or procudes errors.

Signed-off-by: Jan-Philipp Litza <janphilipp@litza.de>