chore(presets): rename `npm:unpublishSafe` to `security:minimumReleaseAgeNpm`
As noted in #37952, we've seen a significant increase in recent years on
a number of package ecosystems being targeted with supply chain security
attacks.

We're looking to introduce a default for `minimumReleaseAge` in the
`best-practices` config preset, which will use the `npm:unpublishSafe`.

However, we'd first like to rename `npm:unpublishSafe` to
`security:minimumReleaseAgeNpm` to fit within the new proposed naming
scheme.

This introduces a config migration to make sure that we can provide
folks a non-breaking path for the new preset name.