fix(gomod): Search the parent directories for "vendor" directory. (#37910)
gomod: Search the parent directories for "vendor" directory.

When the golang workspaces are used, the `go work vendor` generates
the `vendor` and `go.work.sum` files in the same directory where
the `go.work` exists - it means in the parent directory of `go.mod`
files.

The current code does not handle this behaviour correctly - it only
searches for the `vendor` directory next to the `go.mod` file.

In this commit, the code is extended to use `findLocalSiblingOrParent`
function to search for the `vendor` directory the very same way as
we currently search for the `go.work` file.

The code is also extended to search for the `go.work.sum` next to the
`go.work` file.

Related: https://github.com/renovatebot/renovate/discussions/37653

Signed-off-by: Jan Kaluza <jkaluza@redhat.com>
Co-authored-by: Jamie Tanna <jamie.tanna@mend.io>