feat(minimumReleaseAge): optionally enforce presence of `releaseTimestamp`
As noted in #38290, #38348, a "rough edge" of the current implementation
of `minimumReleaseAge` is that:

- if we have set `minimumReleaseAge` to apply to a given dependency
- and we have 3 dependencies that need an update
  - 2 of which have a `releaseTimestamp` that is not yet "stable"
  - 1 of which does not have a `releaseTimestamp`

The dependency update that does not have a `releaseTimestamp` will be
treated as "stable", and will receive an update immediately.

To make this a little more predictable, and explicit, we should add a
new configuration item, `minimumReleaseAgeTimestamp`, to control this
behaviour.

To make this a non-breaking change, we will default it to `optional`,
but make it `required` in our "best practices" configuration.

Because this configuration option could be found in many places, such as
under a `packageRule` or under a manager's specific configuration, in a
similar way to `minimumReleaseAge`, we do not add it as a strongly typed
field to relevant objects.

In the case that `minimumReleaseAgeTimestamp=optional` is being used, we
will WARN on any packages that are being updated, but maybe shouldn't
be, as a way to proactively inform the user of possible misconfiguration.

Closes #38290, #38348.