fix(json-schema): correctly set version in generated JSON Schema (#38621)
fix(tools/json-schema): correctly set version in generated JSON Schema

In #38567 we tried to wire in the version of Renovate into the generated
JSON Schema. This unfortunately didn't work, as we don't have the
release number known at the time we're currently calling the JSON Schema
generation.

To fix this, we can make sure that at the point we're generating the
JSON Schema as part of the release process, we can wire in the version
as part of `generateSchema`.