docs(docker): note digests may have an incorrect `releaseTimestamp`
Until the completion of #38659, we should make it clearer that digests
do not have a `releaseTimestamp` associated with them.

This will also be improved in #38656 to provide the correct
`releaseTimestamp` of a given digest.

As flagged in #38212, this can lead to digests looking newer than they
are, and possibly cause `minimumReleaseAge` not to work.

We should surface this to users, and point to the issue for
documentation purposes.