chore(schema): upgrade to JSON Schema draft-07 (#38751)
As noted in #38749, we need to upgrade to draft-07 of JSON Schema to
upgrade our dependency on AJV.

To do so, we "just" need to update the `$schema`, as we're not using
anything that's been deprecated/removed/changed as part of draft-07.

This doesn't impact anything user-facing, and the schema doesn't change.

This appears to be supported by `vscode-json-languageservice`, which is
used by several IDEs (including VSCode) and editors to perform JSON
Schema validation in-editor.

Note that if we don't update `schemaId` to `auto`, we end up with errors
such as:

    Error: no schema with key or ref "http://json-schema.org/draft-07/schema#"

Closes #38749.