chore(schema): add JSON Schema for global configuration
As noted in #38727, it would be convenient to distribute a separate JSON
Schema that is explicitly for "global" / "admin" / "self-hosted"
configuration, separate to the repository configuration in our current
`renovate-schema.json`.

This gets wired in as an additional JSON file, and will have a different
set of configuration options in the future, noted by the `isGlobal`.

Closes #38727.