chore(schema): deprecate `globalOnly` options in renovate-schema.json (#38805)
Now we have a global schema for self-hosting configuration, we can
deprecate the use of global configuration in a repository configuration.

Unfortunately the JSON Schema version we use does not provide native
deprecation functionality, so we can prepend the deprecation notice to
the `description`.

Then, as part of #38728, we will remove this in the next major version
of Renovate.

Closes #38729.