feat(minimumReleaseAge)!: require a release timestamps by default
As noted in #38290, #38348, a "rough edge" of the current implementation
of `minimumReleaseAge` is that:

- if we have set `minimumReleaseAge` to apply to a given dependency
- and we have 3 dependencies that need an update
  - 2 of which have a `releaseTimestamp` that is not yet "stable"
  - 1 of which does not have a `releaseTimestamp`

The dependency update that does not have a `releaseTimestamp` will be
treated as "stable", and will receive an update immediately.

To provide safer "best practices" we will provide a default value for
`minimumReleaseAgeBehaviour` in `config:best-practices` to protect
users.

Note that this won't change anything unless there are already
dependencies being matched by `minimumReleaseAge`, but is still a
breaking change, as it will impact all users who are - now or in the
future - setting `minimumReleaseAge`.

Closes #38457.