chore: add separate input for version requirement on "request help"
We end up getting quite a few Discussions where the version is indicated
as "latest", which can be hard to determine, as that version changes
fairly regularly.

Sometimes we also get Docker image digest, which are more easy to
determine what they pin to, but still less straightforward than an exact
version.

To save some back-and-forth with confirming with users, let's make this
clearer in the "Request Help" template, by creating a separate "which
platform are you self hosting with" and "what version are you using".