fix(util/fs): ensure that return value of privateCacheDir() is unique
When running in parallel renovatebot instances would use the same value
for `privateCacheDir()` leading to cases where one instance of renovate
could delete the cache from under the feet of another.

This commit provides a MVP solution to this issue by ensuring that
the return value from `privateCacheDir()` is unique for a given run of
renovate.

Due to the way `privateCacheDir()` is used by the rest of renovate, it
is assumed to have a fixed return value and so we memoize the output.
This keeps a small chance of race in that the directory is created
safely once but subsequently is simply removed and re-created which does
slightly push against established best practice for temporary
directories.

A "fuller fix" would involve the repository processing logic calling
`privateCacheDir()` *once* per repository, remembering the output and
deleting the directory once done.