fix(config/validation): validate `allowedValues` for arrays
To make sure that we provide faster feedback for users, we should
validate that for options that have an `allowedValues` set, we validate
the option according to the value.

We also need to:

- handle a single option value
- handle an array of option values
- handle when a default (`null`) option is used, as that won't match
  `allowedValues`
- handle when a `supportsTemplating` config option is used alongside
  `allowedValues`
- add a special case to handle `packageRule.versioning` differently, as
  it may have a `regex:` value which isn't technically in the
  `allowedValues`

This adds a number of extra errors for some validation that is already
happening for specific options' values, which we'll follow-up and remove
the duplication for.

This also makes sure that we validate this for both repository config
and global config.