feat(presets): add global preset to allow unsafe actions
As a means to more easily support self-hosted administrators who wish
to allowlist certain configuration, we can introduce a new preset,
`global:allowAllUnsafe`, to turn the configuration all on.

We can add a warning - both in the description and the name - to hint to
administrators that they should be /very/ careful when enabling this due
to the additional remote code execution this allows.