feat(presets): log the resolved configuration (without internal presets)
As noted in #15827, it would be useful to log the "shallow"
configuration for a given repository - that is, the resolved
configuration, but without any Renovate internal presets being included.

This will be very useful for debugging, as it shows the final resulting
configuration, but without the - often very - verbose internal presets.

In the case that only this log line is shared, we can make sure to
reference the version of the Renovate CLI in the log line.

This requires building on top of changes in
6a4fa3f245dd8976fc8477c4d72f7aae32adab34 to be able to return a `merged`
and `unmerged` list of presets, and uses helpers from
c8f42e86b4f182a6c4c530250b2b1d2a0fb20e22, and
66b84dd12bd9a638b914883fa2781e2def601e9a to process the internal presets
accordingly.

Since we first worked on this, we've worked on improvements to caching
which should provide a require no additional performance overhead
performing this resolution twice.

This can also lead to being able to understand more clearly how Renovate
configuration is being used[0], as we have the final resolved configuration.

[0]: https://www.jvt.me/posts/2024/04/14/renovate-config-sql/