fix(presets): allow `lockFileMaintenance` to not be governed by `minimumReleaseAge`
As noted in #39352, with Renovate 42's default enforcement of all npm
datasource package updates requiring `minimumReleaseAge` checks to
succeed, we're seeing that lock file maintenance updates are also being
blocked by this, despite Renovate itself delegating the logic to the
package manager.

We can opt-out the `lockFileMaintenance` updates from Minimum Release
Age checks, but make sure that we warn users of the risks.