fix(presets): allow `replacement`s to not be governed by `minimumReleaseAge`
Similar to changes in c1469ab79cf3afa178b3e5b9cf1bf4309340a09e, we also
need to make sure that `minimumReleaseAge` checks are not applying to
`replacement`s, given we do not set a release timestamp (#39400) on
replacement suggestions.

This will be in place until #39400, when we should be able to remove
this workaround.