docs(self-hosting): further clarify the supply chain security risks
As part of ongoing efforts I'm taking to make Renovate have more
secure-by-default settings, we should document some of the additional
risks that can occur when self-hosting.

This primarily focusses on outsider attacks (i.e. from a malicious
dependency) as we're finding is increasing in frequency in the recent
years.