fix(presets): don't override `prCreation` on `security:minimumReleaseAgeNpm` (#39536)
Since rolling out the v42 Renovate release to the Mend Developer
Platform, we've been watching out for user reports for issues following
the addition of `security:minimumReleaseAgeNpm` in
`config:best-practices`.

In particular, two Discussions #39525 and #39477 are interesting cases
where a change from #39066 around `prCreation` has inadvertently caused
impact to PRs being created, as the CI needs to pass on the branch
before a PR is created (as intended by `prCreation=not-pending`).

However, these branches do not get CI running until there is a PR, which
means PRs will never be raised.

This partially reverts changes from #39066, #39116 and #38873, which
have been re-investigated since the reports.

Slightly surprising is the fact that the changes from #39066 have been
live on the Mend Developer Platform for a few weeks, but it wasn't
noticed until now, after the full rollout of v42.