docs: add Renovate CLI version to docs footer (#39568)
As part of wiring in the version of the CLI into the docs build process,
we should also show this version to users, so they're aware what version
of the docs they're looking at.

To put this into the footer, the simplest place is to overwrite the
existing `copyright.html`[0], and remove the (unused) copyright
configuration.

We make the inclusion of the version optional, to make it possible to
merge + release this before the docs site has the version wired in.

For this to work, we need to fetch the `RENOVATE_VERSION` from the
environment (if present).

[0]: https://github.com/squidfunk/mkdocs-material/blob/9.7.0/material/templates/partials/copyright.html