fix(instrumentation): correctly instrument `extract` and `lookup`
When calling `extractDependencies`, we actually perform both `extract`
and `lookup` splits. However, we are instrumenting the whole method call
as the `extract` split, which results in incorrect tracing data.

To correct this, we can wrap the specific code responsible for the
extraction into `extract`, and then the separate code for `lookup`
accordingly.

This also makes sure that `emptyExtract` sets repository splits.