fix(instrumentation): correctly `async` / `await` Git operations
A previous refactor had incorrectly made it so we were not correctly
`await`ing any calls to the SimpleGit method calls.

This has mostly led to the Git operations continuing to work, but the
instrumentation hasn't been correctly rendering.

This also corrects incorrect return types that were previously not
returned correctly.