fix(cache): instrument cache creation + destroy
As we're already calculating the duration of operations, we can make
sure that we add our instrumentation around it.

This will then provide instrumentation across each of our cache
implementations.

Note that this doesn't introduce a span for each cache key we're looking
to clean up, given a small repo can have hundreds of cached items, let
alone for large repositories.