fix(instrumentation): handle Promise-like objects as Promises
When instrumenting, we specifically handle Promises. However, if we have
an object, like SimpleGit's `Response`:

```
export type Response<T> = SimpleGit & Promise<T>;
```

This does not get correctly instrumented, and we end up not waiting for
the function call to complete.

We can make sure we use `isPromise` to correctly check if it's a native
Promise, or something with a Promise-like API.