fix(global): ensure that all `globalOnly` options can be set with Global Config
As part of #9603, we intended to come back and make sure that all of our
`globalOnly` configuration fields are set here, so they can be correctly
updated by configuration (in various forms).

As some configuration may not work if we do not set this, we should make
sure that this is applied, as well as ensuring that there is a test to
catch this requirement.