test: don't sign commits made in submodules created by test suite (#39748)
fix: don't sign commits made in submodules created by test suite

The 'commit.gpgsign=false' configuration was already in place for some
tests in this file, but not all of them. I'm pretty sure that with this
change in place, there's no more commit signing happening in this test
suite.