feat: add support for go:generate directives in gomod (#39163)
* feat: add support for go:generate directives in gomod

This change adds a `goGenerate` postUpdateOption which enables
running `go generate ./...` after a dependency update.

Frequently when dependencies are brought in via tools or using
vendored code, and those dependencies are used for go:generate
instructions, we want to update the generated code after the dependency
has been updated.

Adding this as a postUpdateCommand wasn't really satisfactory since that
necessarily requires making a change in two spots (one in the allowed
commands, one in the postUpdateCommand). The other reason this was
undesirable was it would cause generate to run after every dependency
update, regardless of language.

* Updates

* Update fixes

* Self PR

* Fix docs link

* Trying to find the right path

* Fix

* Add deleted

* Fix

* And fix test...

* docs: tweak wording for `goGenerate`

* fix(types): don't default GlobalConfig

In a future change, `GlobalConfig.get` will correctly pull in the
defaults.

---------

Co-authored-by: Jamie Tanna <jamie.tanna@mend.io>