fix(github-actions): handling of quoted Docker image references (#39635)
* fix: handling of quoted Docker image references in Github Actions

The problem with the use of '\s' is that it swallows the newline at the
end of the line, meaning the $ anchor doesn't have anything to match
against, which results in it failing to match quoted entries. By simply
switching to a dot (which doesn't match newlines), the issue is
resolved for all references.

* fix: further modify regex for docker image refs in github actions