fix(Gradle): Limit Gradle daemon heap size
This change introduces a hard-coded heap size for the Gradle daemon used when updating the Gradle wrapper and updating the Gradle dependency lockfile(s).

The heap size is limited to 256m for Gradle wrapper updates and 768m for dependency lockfile(s). The higher value for the latter is to give more complex builds a chance to execute. The wrapper update is pretty simple and doesn't need that much heap.

Also ensured that the Gradle daemon terminates early by adding a `--no-daemon` for the wrapper-update.

Similar for the dependency lockfile(s) added a `--stop` to terminate running daemons, but allow the daemon to stay around and speed up the multiple Gradle invocations.

Fixes #39558