fix(util/exec): allow running with `shell: true` for Containerbase installs
After recent changes to remove `shell: true`, it's clear from #40231
that this breaks Containerbase's installing of tools, as we use the
`hash` shell built-in.

This requires we make sure that the `shell` option is passed around.

We will temporarily re-enable this, and then follow up in #40232 to
remove this requirement.